Long the timestamp.

+++ src/tsdbmigrator/Main.java
@@ -186,7 +186,7 @@ final class Main {
     System.arraycopy(orig_key, SALT_WIDTH + METRICS_WIDTH + TIMESTAMP_BYTES, new_col, TIMESTAMP_BYTES, new_col.length - TIMESTAMP_BYTES);
 
     // TODO - prevent duplicate puts here.
-    mutation.withTimestamp(month * 1000 * 1000) // microseconds
+    mutation.withTimestamp(((long) month) * 1000 * 1000) // microseconds
             .withRow(CassandraClient.TSDB_T_INDEX, new_key).putColumn(new_col, new byte[]{0});
   }
